vulncode/requirements.txt
Flask==0.12
requests==2.19.1
urllib3==1.24.1
Jinja2==2.10.1
PyYAML==5.1
cryptography==2.3